pyproject.toml
[build-system]
requires = ["hatchling>=1.19"]
build-backend = "hatchling.build"

[project]
name = "cornerstone-app"
version = "0.1.0"
description = "Embedding-based search prototype"
requires-python = ">=3.13"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.110,<0.111",
    "uvicorn[standard]>=0.24,<0.30",
    "qdrant-client>=1.7,<2.0",
    "openai>=1.0,<2.0",
    "sentence-transformers>=2.2,<3.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6,<3.0",
    "pyyaml>=6.0,<7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2,<9.0",
    "pytest-asyncio>=0.23,<0.24",
    "httpx>=0.26,<0.28"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cornerstone"]


[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra"
markers = ["integration: tests requiring the Docker-hosted Qdrant instance"]
